In this version of the code, we are once again reading a VDW parameter file with all GAFF and AMBER atom types definition. The file must contain the radius and epsilon values for each atom type found in GAFF and in AMBER. This file will only be required if mol2_aa is set to true.
The file structure must be something like:
  hc          1.4593  0.0208
  ha          1.4735  0.0161
  hn          0.6210  0.0100
  ho          0.3019  0.0047
  hs          0.6112  0.0124
  hp          0.6031  0.0144
  o           1.7107  0.1463
  os          1.7713  0.0726
  oh          1.8200  0.0930
  c3          1.9069  0.1078
  c2          1.8606  0.0988
  c1          1.9525  0.1596
...

Also, the file MUST BE located in ${LIBELA}/src/tools/gaff2_vdw.dat,
otherwise the program will exit returning 1.

This test is designed to evaluate the need of different atom types for
the same element. In Sybyl types, hydrogen atoms are defined with only
one type, namely "H". On the other hand, in GAFF, there are at least 6
different types for hydrogen atoms.