This version has the first version of the fully flexible approach in MC simulations, where ligands can sample every degree of freedom for each atom, without the need of sampling torsion angles.
The simulation is much slower, and does not require the OpenBabel API to
find and sample torsion angles. On the other hand, this API is used to
compute ligand internal energies for each MC trial step.

This approach is used by setting the input flag "mc_full_flex" to "yes".
The algorithm now looks for one of these settings:

(i) generate_conformers? if set to yes, samples among pre-generated
conformers.

(ii) sample_torsions ? if set to yes, samples torsion angles in MC.

(iii) mc_full_flex ? if set to yes, samples every dof for atoms;

(iv) if none of the above are set to yes, it will do rigid body sampling
of the ligand in the active site.

Still under testing.