We have an issue with the OpenMp execution for docking.
It seems that sometimes many threads try to write in the same file at
the same time. It may result in some segfaults. Using pragma omp
critical did not help, in a first attempt. I will have to go deeper in
this issue.

For now, the best way to use many threads safely is using MPI.