In this version, I added:
* Support to scale the Electrostatic and VDW terms during docking with
the keywords scale_elec_energy and scale_vdw_energy. These values
(double) are multiplied to the computed energy terms;

* Support for this above mentioned scaling in the GUI;

* Extended info in mol2 files written by docking. Now, the energy terms
are also witten in a format readable by UCSF Chimera, so that the user
can sort molecules for the most favorable in elec/vdw/restraint/solv
terms as well as total energy or similarity index (named as RMSD).