Added a overlay_cutoff option.
In this version, the user can use the following keywords to decide
whether the energy optimization should or should not be done on the
basis of the similarity index computed after the overlay optimization.

For this purpose use, for example:

use_overlay_cutoff    yes
overlay_cutoff
0.60

In this case, molecules with SI >= 0.60 will be used for energy
optimization. Molecules with SI < 0.6 won't be energy optimized and will
be written in the log file (and mol2 file, if asked to) with the correct
similarity index and interaction energy = 0.0;