Major modification to get a working pipeline of dock+eq
This version brings a major modification. We now have a first working
version (tests needed) of a pipeline of docking+eq or docking+mc. The
intention is to setup a post-docking refinment of poses with rapid MC
simulations.

This pipeline is not expected to work with OpenMP due to the lack of
thread-safety in many OpenBabel functions. It might work using some
"pragma critical", but I did not try to implement it yet.

It seems to be working with single execution and with MPI
paralelization.

Changes include:

* The output file names have been changed.