Corrected McGrid class.
The FindHB only parses protein donnors/acceptors. It is not aware of
cofactors in the protein (like ATP, ADP, GDP, GTP, NAD(H), etc). This
restriction is added in the current version and the reason for this is
that it crashes in some cases where a cofactor has an N atom for example
that could be parsed as if it was and protein mainchain N atom.