In this version, we use the overlay_cuttoff parameter to define whether the energy optimization will be perfomed.
From this version on, the energy optimization will only be performed if
there is any conformer with SI > overlay_cuttoff. This saves time in the
docking procedure.

By default, the overlay_cutoff is set as 0.0, i.e., every docking pose
found in the overlay optimization will be scored and enegy optimized.
However, it is up to the user to define his criteria.

In a search in the DUD38 database, we found that the average SI for
actives (across all the targets) is about 0.69+-0.06. So, it seems
reasonable to set the overlay cutoff to something between 0.65 to 0.70,
depending on how restrictive the user wants to be. However, it still
needs extensive testing.