Step back!
It looks like the inclusion of the firt iteration (equilibration
iteration) during the computation of the free energy in MCR simulations
led to an important overestimate of the free energy. Data was taken from
T4L, CD and CB7 datasets. Taken together, we see an slope of about 5
when plotting experimental vs calculated free energies. The r^2 is not
that bad and, of course, linearity is preserved.

So we are taking a step back here. And starting the MCR simulations with
a equilibration simulation (bi=2.0 and not using W for dG calculation)
prior to actual recursion over the given coefficients.